Separate IPFIX templates into separate file
